PDF renderer overhaul
A new configuration option 'renderer.renderer_class' has been
specified so that it is now possible to specify the renderer class to be
used. The 'output_file' configuration option has been moved into the
reporter module.

On the other hand the PDF renderer has been refactored, and now it is
not needed to create any intermediate files for generating the PDF file.