Use loadables for collectors
Since we are going to have several collectors in the near future we need
a mechanism to automatically load them, and execute them as appropriate
to get the defined metrics. This introduces a new list option "collectors"
that is to be set to the class names of the collectors being used.